+++ Source/DynamicScene/DynamicPolylineVisualizer.js
@@ -4,7 +4,7 @@ define([
         '../Core/defined',
         '../Core/destroyObject',
         '../Core/Cartesian3',
-        '../Core/Color',
+        './MaterialProperty',
         '../Scene/Material',
         '../Scene/PolylineCollection'
        ], function(
@@ -12,7 +12,7 @@ define([
          defined,
          destroyObject,
          Cartesian3,
-         Color,
+         MaterialProperty,
          Material,
          PolylineCollection) {
     "use strict";
@@ -185,6 +185,7 @@ define([
         var vertexPositionsProperty = dynamicObject._vertexPositions;
         var polylineVisualizerIndex = dynamicObject._polylineVisualizerIndex;
         var show = dynamicObject.isAvailable(time) && (!defined(showProperty) || showProperty.getValue(time));
+        var context = dynamicPolylineVisualizer._scene.getContext();
 
         if (!show || //
            (!defined(vertexPositionsProperty) && //
@@ -199,7 +200,6 @@ define([
             return;
         }
 
-        var uniforms;
         if (!defined(polylineVisualizerIndex)) {
             var unusedIndexes = dynamicPolylineVisualizer._unusedIndexes;
             var length = unusedIndexes.length;
@@ -220,13 +220,8 @@ define([
                 material = Material.fromType(Material.PolylineOutlineType);
                 polyline.setMaterial(material);
             }
-            uniforms = material.uniforms;
-            Color.clone(Color.WHITE, uniforms.color);
-            Color.clone(Color.BLACK, uniforms.outlineColor);
-            uniforms.outlineWidth = 0;
         } else {
             polyline = dynamicPolylineVisualizer._polylineCollection.get(polylineVisualizerIndex);
-            uniforms = polyline.getMaterial().uniforms;
         }
 
         polyline.setShow(true);
@@ -243,19 +238,9 @@ define([
             polyline._visualizerPositions = vertexPositions;
         }
 
-        var property = dynamicPolyline._color;
+        var property = dynamicPolyline._material;
         if (defined(property)) {
-            uniforms.color = property.getValue(time, uniforms.color);
-        }
-
-        property = dynamicPolyline._outlineColor;
-        if (defined(property)) {
-            uniforms.outlineColor = property.getValue(time, uniforms.outlineColor);
-        }
-
-        property = dynamicPolyline._outlineWidth;
-        if (defined(property)) {
-            uniforms.outlineWidth = property.getValue(time);
+            polyline.setMaterial(MaterialProperty.getValue(time, property, polyline.getMaterial()));
         }
 
         property = dynamicPolyline._width;